move reducers for poster
