+++ packages/space-opera/src/redux/space_opera_base_test.ts
@@ -21,9 +21,9 @@ import {cloneJson} from '@google/model-viewer-editing-adapter/lib/util/clone_jso
 import {RGBA} from '@google/model-viewer/lib/model-viewer';
 
 import {dispatchAddBaseColorTexture, dispatchAddEmissiveTexture, dispatchAddMetallicRoughnessTexture, dispatchAddNormalTexture, dispatchAddOcclusionTexture, dispatchBaseColorTexture, dispatchEmissiveTexture, dispatchMaterialBaseColor, dispatchMetallicFactor, dispatchNormalTexture, dispatchOcclusionTexture, dispatchRoughnessFactor, dispatchSetAlphaCutoff, dispatchSetAlphaMode, dispatchSetEmissiveFactor} from '../components/materials_panel/reducer.js';
+import {applyEdits, generateTextureId, getGltfEdits, INITIAL_GLTF_EDITS} from '../components/model_viewer_preview/gltf_edits.js';
 
 import {applyCameraEdits, Camera} from './camera_state.js';
-import {applyEdits, generateTextureId, getGltfEdits, INITIAL_GLTF_EDITS} from './gltf_edits.js';
 import {dispatchCurrentCameraState, dispatchGltfAndEdits, dispatchGltfUrl, dispatchInitialCameraState, reduxStore, registerStateMutator} from './space_opera_base.js';
 
 const EXAMPLE_BIN_AS_STRING = 'example of some bin data';
